Initial seperation of create and clone with client refactor
- Seperate clone and create into their own steps
- Clean up code a little for betting reading and usability
- Refactor client so it isn't so cluttered
