+++ README.md
@@ -38,7 +38,8 @@ The most basic json file you can build from is:
   "builders": [
     {
       "installer_app": "/Applications/Install macOS Big Sur.app",
-      "type": "veertu-anka"
+      "type": "veertu-anka-vm-create",
+      "vm_name": "macos-big-sur"
     }
   ],
   "post-processors": [
@@ -69,11 +70,12 @@ You can also skip the creation of the base VM template and use an existing VM te
     }
   ],
   "builders": [{
-    "type": "veertu-anka",
+    "type": "veertu-anka-vm-clone",
     "cpu_count": 8,
     "ram_size": "10G",
     "disk_size": "150G",
-    "source_vm_name": "10.15.6"
+    "source_vm_name": "10.15.6",
+    "vm_name": "macos-from-source_10.15.6"
   }]
 }
 ```
@@ -88,15 +90,31 @@ This will clone `10.15.6` to a new VM and, if there are differences from the bas
 
 ## Builders 
 
-### veertu-anka-vm
+### veertu-anka-vm-create
 
 #### Required Configuration
 
+* `installer_app` (String)
+
+The path to a macOS installer. This must be provided if `source_vm_name` isn't provided. This process takes about 20 minutes. The resulting VM template name will be `anka-packer-base-{macOSVersion}`.
+
 * `type` (String)
 
-Must be `veertu-anka-vm`.
+Must be `veertu-anka-vm-create`.
+
+* `vm_name` (String)
+
+The name for the VM that is created. One is generated if not provided (`anka-packer-{10RandomCharacters}`).
 
 #### Optional Configuration
+
+* `anka_password` (String)
+
+Sets the password for the vm. Defaults to `admin`.
+
+* `anka_user` (String)
+
+Sets the username for the vm. Defaults to `anka`.
 
 * `boot_delay` (String)
 
@@ -116,17 +134,13 @@ The size in "[0-9]+G" format, defaults to `25G`.
 
 The Hardware UUID you wish to set (usually generated with `uuidgen`).
 
-* `installer_app` (String)
-
-The path to a macOS installer. This must be provided if `source_vm_name` isn't provided. This process takes about 20 minutes. The resulting VM template name will be `anka-packer-base-{macOSVersion}`.
-
 * `port_forwarding_rules` (Struct)
 
 > If port forwarding rules are already set and you want to not have them fail the packer build, use `packer build --force`
 
 ```json
   "builders": [{
-    "type": "veertu-anka",
+    "type": "veertu-anka-vm-clone",
     "cpu_count": 8,
     "ram_size": "10G",
     "source_vm_name": "anka-packer-base-10.15.7",
@@ -139,7 +153,8 @@ The path to a macOS installer. This must be provided if `source_vm_name` isn't p
       {
         "port_forwarding_guest_port": 8080
       }
-    ]
+    ],
+    "vm_name": "macos-from-packer-base_10.15.7"
   }]
 ```
 
@@ -147,13 +162,89 @@ The path to a macOS installer. This must be provided if `source_vm_name` isn't p
 
 The size in "[0-9]+G" format, defaults to `2G`.
 
+* `stop_vm` (Boolean)
+
+Whether or not to stop the vm after it has been created, defaults to false.
+
+### veertu-anka-vm-clone
+
+#### Required Configuration
+
 * `source_vm_name` (String)
 
 The VM to clone for provisioning, either stopped or suspended.
 
+* `type` (String)
+
+Must be `veertu-anka-vm-clone`.
+
 * `vm_name` (String)
 
 The name for the VM that is created. One is generated if not provided (`anka-packer-{10RandomCharacters}`).
+
+#### Optional Configuration
+
+* `always_fetch` (Boolean)
+
+Always pull the source VM from the registry. Defaults to false.
+
+* `boot_delay` (String)
+
+The time to wait before running packer provisioner commands, defaults to `10s`.
+
+* `cpu_count` (String)
+
+The number of CPU cores, defaults to `2`.
+
+* `disk_size` (String)
+
+The size in "[0-9]+G" format, defaults to `25G`.
+
+> We will automatically resize the internal disk for you by executing: `diskutil apfs resizeContainer disk1 0`
+
+* `hw_uuid` (String)
+
+The Hardware UUID you wish to set (usually generated with `uuidgen`).
+
+* `port_forwarding_rules` (Struct)
+
+> If port forwarding rules are already set and you want to not have them fail the packer build, use `packer build --force`
+
+```json
+  "builders": [{
+    "type": "veertu-anka-vm-clone",
+    "cpu_count": 8,
+    "ram_size": "10G",
+    "source_vm_name": "anka-packer-base-10.15.7",
+    "port_forwarding_rules": [
+      {
+        "port_forwarding_guest_port": 80,
+        "port_forwarding_host_port": 12345,
+        "port_forwarding_rule_name": "website"
+      },
+      {
+        "port_forwarding_guest_port": 8080
+      }
+    ],
+    "vm_name": "macos-from-packer-base_10.15.7"
+  }]
+```
+
+* `ram_size` (String)
+
+The size in "[0-9]+G" format, defaults to `2G`.
+
+* `source_vm_tag` (String)
+
+Specify the tag of the VM we want to clone instead of using the default.
+
+* `stop_vm` (Boolean)
+
+Whether or not to stop the vm after it has been created, defaults to false.
+
+* `update_addons` (Boolean)
+
+Update the vm addons. Defaults to false.
 
 ## Post Processors
