+++ builder/anka/artifact.go
@@ -10,33 +10,33 @@ type Artifact struct {
 	vmId   string
 }
 
-// BuilderId returns the builder Id.
+// BuilderId returns the unique builder id.
 func (*Artifact) BuilderId() string {
 	return BuilderId
 }
 
 // Destroy destroys the image represented by the artifact.
-func (self *Artifact) Destroy() error {
-	// log.Printf("Destroying image: %s", self.String())
-	// err := self.client.destroyImage(self.imageId)
+func (a *Artifact) Destroy() error {
 	return errors.New("Destroy not implemented")
 }
 
 // Files returns the files represented by the artifact.
-func (*Artifact) Files() []string {
+func (a *Artifact) Files() []string {
 	return nil
 }
 
 // Id returns the VM UUID.
-func (self *Artifact) Id() string {
-	return self.vmId
+func (a *Artifact) Id() string {
+	return a.vmId
 }
 
-func (self *Artifact) State(name string) interface{} {
+// State allows the caller to ask for builder specific state information
+// relating to the artifact instance.
+func (a *Artifact) State(name string) interface{} {
 	return nil
 }
 
 // String returns the string representation of the artifact.
-func (self *Artifact) String() string {
-	return self.vmName
+func (a *Artifact) String() string {
+	return a.vmName
 }
