+++ examples/clone-existing.json
@@ -1,6 +1,6 @@
 {
   "variables": {
-    "source_vm_name": ""
+    "source_vm_name": "macos-big-sur"
   },
   "provisioners": [
     {
@@ -13,10 +13,11 @@
     }
   ],
   "builders": [{
-    "type": "veertu-anka-vm",
+    "type": "veertu-anka-vm-clone",
     "cpu_count": 8,
     "ram_size": "10G",
     "disk_size": "150G",
-    "source_vm_name": "{{user `source_vm_name`}}"
+    "source_vm_name": "{{user `source_vm_name`}}",
+    "vm_name": "macos-from-{{user `source_vm_name`}}"
   }]
 }
